7.1.15 Splits selectors
