+++ src/store/data/selectors.js
@@ -1,11 +1,9 @@
 import {createSelector} from "reselect";
-import {NameSpace} from "./store/root-reducer";
-import {sortOffers} from "./utils";
+import {NameSpace} from "../root-reducer";
+import {sortOffers} from "../../utils";
+import {getCities, getCity, getActiveSorter} from "../app/selectors";
 
 const getOffers = (state) => state[NameSpace.DATA].offers;
-const getCity = (state) => state[NameSpace.APP].city;
-const getCities = (cities) => Object.values(cities);
-const getActiveSorter = (state) => state[NameSpace.APP].activeSorter;
 
 const getOffersByCity = createSelector([getOffers, getCity],
     (offers, city) => offers.filter((it) => it.city.name === city));
